golf-pipeline/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "golf-pipeline"
version = "0.1.0"
description = "Egocentric golf swing capture, pose extraction, and metrics pipeline"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # orchestration
    "temporalio>=1.6",

    # gpu inference
    "modal>=0.62",

    # api
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",

    # data
    "pydantic>=2.6",
    "pymongo>=4.6",
    "motor>=3.4",
    "numpy>=1.26",
    "scipy>=1.12",

    # storage
    "boto3>=1.34",

    # video / audio
    "opencv-python-headless>=4.9",
    "librosa>=0.10",
    "soundfile>=0.12",
    "ffmpeg-python>=0.2",

    # pose (cpu fallback for local dev — gpu version runs on modal)
    "mediapipe>=0.10",

    # utilities
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "structlog>=24.1",
    "typer>=0.9",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
golf = "golf_pipeline.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
